２Fシーンの修正
