+++ Assets/Spricts/Manager2F.cs.meta
